Clippy: simplify idioms (sort_by_key, is_empty, map, etc.)
Machine-applicable style/complexity fixes plus two hand-corrected ones:

- `sort_by(|a,b| k(a).cmp(&k(b)))` -> `sort_by_key(|a| k(a))`.
- `.clone()` on `Copy` types (`Cursor`, `LeafData`) -> plain copy.
- `let x = expr; x` -> return `expr` directly.
- `map_or(true, ..)` -> `is_none_or(..)`; `buffer.len() == 0` ->
  `is_empty()`; `saturating_sub(..).max(0)` on usize -> drop no-op
  `.max(0)`; `as u64` on a `u64` -> drop cast.
- `.filter(..).next_back()` -> `.rfind(..)`.
- `if let Some(..) { Some(..) } else { None }` -> `.map(..)`.
- cmp_owned: `name.to_string() == current_map` -> `*name == current_map`
  (hand-written; clippy's auto-suggestion flipped operands and didn't
  typecheck).

Behavior-preserving. `cargo build` clean.

+++ crates/fresh-editor/src/app/plugin_dispatch.rs
@@ -4014,7 +4014,7 @@ impl Editor {
                 // Synthetic, window-derived request id (well clear of the
                 // low-numbered JS callback ids) so the in-flight/cancel
                 // tracking works and a repeated switch doesn't double-connect.
-                let request_id = u64::MAX - window_id.0 as u64;
+                let request_id = u64::MAX - window_id.0;
                 if self.remote_attach_inflight.contains(&request_id) {
                     return;
                 }
